implement REST

+++ app/controllers/comments_controller.rb
@@ -1,43 +1,48 @@
 class CommentsController < ApplicationController
-  before_filter :cur_user
+  before_filter :cur_user , :check_right
 
-  
-  #get 'briefs/brief_id/comments/new' => :new_brief_comment,
-  #  :as => 'new_brief_comment'
-  def new_brief_comment
-    @brief = Brief.find(params[:brief_id])
-    @brief.can_comment?(@cur_user)
-    
-    @comment = Comment.new
-    @back = params[:back]
-    @path = create_brief_comment_path(@brief,:back => @back) 
+  def check_right
+    case @cur_user.org
+    when RpmOrg , CheilOrg
+    else  raise SecurityError
+    end
   end
 
-  #post 'briefs/:brief_id/comments' => :create_brief_comment,
-  #  :as => 'create_brief_comment'
-  def create_brief_comment
-    brief = Brief.find(params[:brief_id])
-    brief.can_comment?(@cur_user)
-    
-    comment = BriefComment.new(params[:comment])
-    comment.user_id = @cur_user.id
-    brief.comments << comment
+  #get 'briefs/brief_id/comments/new' => :new,
+  #  :as => 'new_brief_comment'
+  def new
+    if params[:brief_id]
+      @brief = Brief.find(params[:brief_id])
+      @brief.check_comment_right(@cur_user)
+      @comment = Comment.new
+      @back=params[:back]
+      @path = brief_comments_path(@brief,:back=>@back)
+    end
+  end
 
+  #post 'briefs/:brief_id/comments' => :create,
+  #  :as => 'brief_comments'
+  def create
+    if params[:brief_id]
+      brief = Brief.find(params[:brief_id])
+      brief.check_comment_right(@cur_user)
+
+      comment = BriefComment.new(params[:comment])
+      comment.user_id = @cur_user.id
+      brief.comments << comment
+    end
     redirect_to params[:back]
   end
 
-  #delete 'briefs/:brief_id/comments/:comment_id' => :destroy_brief_comment,
-  #  :as => 'destroy_brief_comment' 
-  def destroy_brief_comment
-    brief = Brief.find(params[:brief_id])
-    brief.can_comment?(@cur_user)
-
-    comment = brief.comments.find(params[:comment_id])
-    comment.destroy
-
-    respond_to do |format|
-      format.html { redirect_to params[:back] }
-      format.json { head :ok }
+  #delete 'briefs/:brief_id/comments/:comment_id' => :destroy,
+  #  :as => 'brief_comment' 
+  def destroy
+    if params[:brief_id]
+      brief = Brief.find(params[:brief_id])
+      comment = brief.comments.find(params[:id])
+      comment.check_destroy_right(@cur_user)
+      comment.destroy
     end
+    redirect_to params[:back] 
   end
 end
